Deciding to use a new path
The text here mentioned the decision to use a new path twice.  Also,
probing packets almost always include padding frames now: we require
that the datagram is full-sized when probing so that we can probe for
MTU, and probing is almost certainly done with short header packets,
where padding has to be inside the packet.